Add selector and type files for useLocalMedia
