+++ src/lib/core/redux/slices/__tests__/rtcConnection.unit.ts
@@ -5,7 +5,7 @@ import {
     selectStreamsToAccept,
 } from "../rtcConnection";
 import { oneOf, randomRemoteParticipant } from "../../../../../lib/__mocks__/appMocks";
-import { StreamState } from "~/lib/RoomParticipant";
+import { StreamState } from "../../../../../lib/RoomParticipant";
 
 describe("rtcConnectionSlice", () => {
     describe("reactors", () => {
