+++ src/lib/react/useRoomConnection/types.ts
@@ -2,7 +2,7 @@ import { LocalParticipant, RemoteParticipant, Screenshare } from "../../RoomPart
 
 import { ChatMessage as SignalChatMessage } from "@whereby/jslib-media/src/utils/ServerSocket";
 import { LocalMediaOptions } from "../../core/redux/slices/localMedia";
-import { UseLocalMediaResult } from "../useLocalMedia";
+import { UseLocalMediaResult } from "../useLocalMedia/types";
 
 export type RemoteParticipantState = Omit<RemoteParticipant, "newJoiner" | "streams">;
 export type LocalParticipantState = LocalParticipant;
@@ -59,4 +59,22 @@ export interface RoomConnectionOptions {
     roomKey?: string;
     localMedia?: UseLocalMediaResult;
     externalId?: string;
+}
+
+export interface UseRoomConnectionOptions extends Omit<RoomConnectionOptions, "localMedia"> {
+    localMedia?: UseLocalMediaResult;
+}
+
+export interface RoomConnectionActions {
+    sendChatMessage(text: string): void;
+    knock(): void;
+    setDisplayName(displayName: string): void;
+    toggleCamera(enabled?: boolean): void;
+    toggleMicrophone(enabled?: boolean): void;
+    acceptWaitingParticipant(participantId: string): void;
+    rejectWaitingParticipant(participantId: string): void;
+    startCloudRecording(): void;
+    startScreenshare(): void;
+    stopCloudRecording(): void;
+    stopScreenshare(): void;
 }
